initial implementation of checksum calculation during deserialization

+++ tools/src/main/java/com/nchain/jcl/tools/bytes/ByteArray.java
@@ -17,7 +17,7 @@ public interface ByteArray {
     boolean isEmpty();
     void add(byte[] data);
     void add(byte[] data, int offset, int length);
-    byte[] get(int offset, int length);
+    byte[] get(long offset, int length);
     byte[] get(int length);
     byte[] get();
     byte[] extract(int length);
